requirements.txt
PyQt5==5.15.10
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
cryptography==43.0.1
email-validator==2.2.0
pyinstaller==6.10.0
keyring==25.3.0
PyQtChart==5.15.6
